% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blaze_version}
\alias{blaze_version}
\title{The version of Blaze used in RcppBlaze}
\usage{
blaze_version(single)
}
\arguments{
\item{single}{A logical value indicates which type to return. If TRUE, it returns an integer. If FALSE, it returns a named vector.}
}
\value{
A number or a named vector to represent the version of \code{blaze} depending on the input, \code{single}.
}
\description{
To return the version of Blaze used in RcppBlaze.
}
\examples{
blaze_version(FALSE)
}
\seealso{
Blaze header file \code{blaze/system/Version.h}.
}
