% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blaze_get_threads}
\alias{blaze_get_threads}
\alias{blaze_set_threads}
\title{Get/Set the Number of Threads used in blaze}
\usage{
blaze_get_threads()

blaze_set_threads(n)
}
\arguments{
\item{n}{The number of threads to set in blaze.}
}
\value{
\code{blaze_get_threads} returns an integer and \code{blaze_set_threads} returns nothing.
}
\description{
Get/Set the Number of Threads used in blaze
}
\seealso{
blaze wiki: \url{https://bitbucket.org/blaze-lib/blaze/wiki/Shared\%20Memory\%20Parallelization}.
}
