% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blaze_set_num_threads}
\alias{blaze_set_num_threads}
\alias{blaze_get_num_threads}
\title{Set/Get the Number of Threads used in blaze}
\usage{
blaze_set_num_threads(n)

blaze_get_num_threads()
}
\arguments{
\item{n}{The number of threads to set in blaze.}
}
\value{
\code{blaze_get_threads} returns an integer and \code{blaze_set_threads} returns nothing.
}
\description{
Set/Get the Number of Threads used in blaze
}
\seealso{
blaze wiki: \url{https://bitbucket.org/blaze-lib/blaze/wiki/Shared\%20Memory\%20Parallelization}.
}
